[ConstitutiveLawsApplication] Rename integrators to avoid Windows 10 error
