+++ applications/ConstitutiveLawsApplication/tests/cpp_tests/test_high_cycle_fatigue_constitutive_law_3d.cpp
@@ -21,7 +21,7 @@
 
 // Application includes
 // Integrator
-#include "custom_constitutive/auxiliar_files/constitutive_laws_integrators/generic_constitutive_law_integrator_damage.h"
+#include "custom_constitutive/auxiliar_files/constitutive_laws_integrators/generic_cl_integrator_damage.h"
 // Yield surfaces
 #include "custom_constitutive/auxiliar_files/yield_surfaces/von_mises_yield_surface.h"
 #include "custom_constitutive/auxiliar_files/plastic_potentials/von_mises_plastic_potential.h"
